refactor: remove streaming support
Remove all streaming/session intent support from the Python SDK:

- Delete src/mpay/methods/tempo/stream/ directory (chain, errors, receipt,
  storage, types, voucher)
- Delete tests/test_stream.py
- Remove StreamMethod class and StreamIntent class
- Remove Mpay.stream() server method
- Remove all streaming imports from __init__.py, client.py, intents.py
- Extract get_tx_params into _rpc.py (was previously in stream/chain.py,
  but needed by TempoMethod for charge transactions)
