+++ src/mpay/methods/tempo/client.py
@@ -1,24 +1,16 @@
 """Tempo payment methods for client-side credential creation.
 
-Implements charge (TempoMethod) and stream (StreamMethod) client methods.
+Implements the charge (TempoMethod) client method.
 """
 
 from __future__ import annotations
 
-import secrets
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from mpay import Challenge, Credential
 from mpay.methods.tempo._defaults import RPC_URL
-from mpay.methods.tempo.stream.chain import (
-    compute_channel_id,
-    encode_approve_call,
-    encode_open_call,
-    get_on_chain_channel,
-    get_tx_params,
-)
-from mpay.methods.tempo.stream.voucher import sign_voucher
+from mpay.methods.tempo._rpc import get_tx_params
 
 if TYPE_CHECKING:
     from mpay.methods.tempo.account import TempoAccount
@@ -26,7 +18,6 @@ if TYPE_CHECKING:
 
 
 DEFAULT_GAS_LIMIT = 100_000
-STREAM_GAS_LIMIT = 2_000_000
 
 
 class TransactionError(Exception):
@@ -195,428 +186,6 @@ class TempoMethod:
 
 
 # ──────────────────────────────────────────────────────────────────
-# Stream client method
-# ──────────────────────────────────────────────────────────────────
-
-
-def _random_salt() -> str:
-    """Generate a random 32-byte salt as 0x-prefixed hex."""
-    return "0x" + secrets.token_hex(32)
-
-
-@dataclass
-class _ChannelEntry:
-    """Tracks a locally-managed channel."""
-
-    channel_id: str
-    salt: str
-    cumulative_amount: int
-    opened: bool
-
-
-@dataclass
-class StreamMethod:
-    """Client-side stream payment method.
-
-    Supports two modes:
-
-    - **Auto mode** (``deposit`` set): Manages the full channel lifecycle
-      (open, incremental vouchers, channel recovery) automatically.
-    - **Manual mode** (``context.action`` set): Caller provides explicit
-      action parameters (open, topUp, voucher, close).
-
-    Example (auto mode)::
-
-        from mpay.methods.tempo import TempoAccount, StreamMethod
-
-        account = TempoAccount.from_key("0x...")
-        method = StreamMethod(
-            account=account,
-            deposit=10_000_000,
-        )
-
-    Example (manual mode)::
-
-        method = StreamMethod(account=account)
-        # Then pass context with action/channelId/etc. to create_credential
-    """
-
-    name: str = "tempo"
-    account: TempoAccount | None = None
-    rpc_url: str = RPC_URL
-    deposit: int | None = None
-    escrow_contract: str | None = None
-    currency: str | None = None
-    recipient: str | None = None
-    decimals: int = 6
-
-    _channels: dict[str, _ChannelEntry] = field(default_factory=dict)
-    _escrow_map: dict[str, str] = field(default_factory=dict)
-
-    @property
-    def intents(self) -> dict[str, Any]:
-        """Available intents -- stream only."""
-        return {}
-
-    async def create_credential(
-        self,
-        challenge: Challenge,
-        context: dict[str, Any] | None = None,
-    ) -> Credential:
-        """Create a stream credential for the given challenge.
-
-        Args:
-            challenge: Payment challenge from the server.
-            context: Optional manual-mode context with action, channelId, etc.
-
-        Returns:
-            A Credential with the stream payload.
-        """
-        if self.account is None:
-            raise ValueError("No account configured for signing")
-
-        account = self.account
-
-        if context and context.get("action"):
-            payload = await self._manual_credential(challenge, account, context)
-        elif self.deposit is not None:
-            payload = await self._auto_manage_credential(challenge, account)
-        else:
-            raise ValueError(
-                "No action in context and no deposit configured. "
-                "Either provide context with action/channelId/"
-                "cumulativeAmount, "
-                "or configure deposit for auto-management."
-            )
-
-        chain_id = self._resolve_chain_id(challenge)
-        return Credential(
-            challenge=challenge.to_echo(),
-            payload=payload,
-            source=f"did:pkh:eip155:{chain_id}:{account.address}",
-        )
-
-    # ──────────────────────────────────────────────────────────
-    # Auto-management mode
-    # ──────────────────────────────────────────────────────────
-
-    async def _auto_manage_credential(
-        self,
-        challenge: Challenge,
-        account: TempoAccount,
-    ) -> dict[str, Any]:
-        """Auto-manage channel lifecycle: open, cumulative vouchers."""
-        md = self._get_method_details(challenge)
-        chain_id = md.get("chainId", 0)
-        escrow = self._resolve_escrow(challenge, chain_id)
-        payee = challenge.request.get("recipient", "")
-        currency = challenge.request.get("currency", "")
-        amount = int(challenge.request.get("amount", "0"))
-        deposit = self.deposit
-
-        if deposit is None:
-            raise ValueError("deposit must be set for auto-management")
-
-        key = self._channel_key(payee, currency, escrow)
-        entry = self._channels.get(key)
-
-        if entry is None:
-            suggested = md.get("channelId")
-            if suggested:
-                entry = await self._try_recover_channel(escrow, suggested, key)
-
-        if entry is not None and entry.opened:
-            entry.cumulative_amount += amount
-            payload = await self._voucher_payload(
-                account,
-                entry.channel_id,
-                entry.cumulative_amount,
-                escrow,
-                chain_id,
-            )
-        else:
-            entry, payload = await self._open_channel(
-                account,
-                escrow,
-                payee,
-                currency,
-                deposit,
-                amount,
-                chain_id,
-            )
-            self._channels[key] = entry
-            self._escrow_map[entry.channel_id] = escrow
-
-        return payload
-
-    async def _open_channel(
-        self,
-        account: TempoAccount,
-        escrow_contract: str,
-        payee: str,
-        currency: str,
-        deposit: int,
-        initial_amount: int,
-        chain_id: int,
-    ) -> tuple[_ChannelEntry, dict[str, Any]]:
-        """Open a new payment channel."""
-        from pytempo import Call, TempoTransaction
-
-        salt = _random_salt()
-
-        channel_id = await compute_channel_id(
-            rpc_url=self.rpc_url,
-            escrow_contract=escrow_contract,
-            payer=account.address,
-            payee=payee,
-            token=currency,
-            deposit=deposit,
-            salt=salt,
-            authorized_signer=account.address,
-        )
-
-        approve_data = encode_approve_call(escrow_contract, deposit)
-        open_data = encode_open_call(payee, currency, deposit, salt, account.address)
-
-        chain_id_val, nonce, gas_price = await get_tx_params(self.rpc_url, account.address)
-
-        tx = TempoTransaction.create(
-            chain_id=chain_id_val,
-            gas_limit=STREAM_GAS_LIMIT,
-            max_fee_per_gas=gas_price,
-            max_priority_fee_per_gas=gas_price,
-            nonce=nonce,
-            nonce_key=0,
-            fee_token=currency,
-            calls=(
-                Call.create(to=currency, value=0, data=approve_data),
-                Call.create(to=escrow_contract, value=0, data=open_data),
-            ),
-        )
-        signed = tx.sign(account.private_key)
-        transaction = "0x" + signed.encode().hex()
-
-        from mpay.methods.tempo.stream.types import Voucher
-
-        voucher = Voucher(channel_id=channel_id, cumulative_amount=initial_amount)
-        signature = sign_voucher(account, voucher, escrow_contract, chain_id)
-
-        entry = _ChannelEntry(
-            channel_id=channel_id,
-            salt=salt,
-            cumulative_amount=initial_amount,
-            opened=True,
-        )
-
-        payload: dict[str, Any] = {
-            "action": "open",
-            "type": "transaction",
-            "channelId": channel_id,
-            "transaction": transaction,
-            "authorizedSigner": account.address,
-            "cumulativeAmount": str(initial_amount),
-            "signature": signature,
-        }
-
-        return entry, payload
-
-    async def _try_recover_channel(
-        self,
-        escrow_contract: str,
-        channel_id: str,
-        key: str,
-    ) -> _ChannelEntry | None:
-        """Attempt to recover an existing channel from on-chain state."""
-        import logging
-
-        try:
-            on_chain = await get_on_chain_channel(self.rpc_url, escrow_contract, channel_id)
-            if on_chain.deposit > 0 and not on_chain.finalized:
-                entry = _ChannelEntry(
-                    channel_id=channel_id,
-                    salt="0x",
-                    cumulative_amount=on_chain.settled,
-                    opened=True,
-                )
-                self._channels[key] = entry
-                self._escrow_map[channel_id] = escrow_contract
-                return entry
-        except Exception:
-            logging.getLogger(__name__).debug(
-                "channel recovery failed for %s, will open new channel",
-                channel_id,
-                exc_info=True,
-            )
-        return None
-
-    # ──────────────────────────────────────────────────────────
-    # Manual mode
-    # ──────────────────────────────────────────────────────────
-
-    async def _manual_credential(
-        self,
-        challenge: Challenge,
-        account: TempoAccount,
-        context: dict[str, Any],
-    ) -> dict[str, Any]:
-        """Build credential payload for a manually-specified action."""
-        md = self._get_method_details(challenge)
-        chain_id = md.get("chainId", 0)
-        channel_id = context.get("channelId", "")
-        escrow = self._resolve_escrow(challenge, chain_id, channel_id)
-        self._escrow_map[channel_id] = escrow
-
-        action = context["action"]
-
-        if action == "open":
-            transaction = context.get("transaction")
-            cumulative_amount = context.get("cumulativeAmount")
-            if not transaction:
-                raise ValueError("transaction required for open action")
-            if cumulative_amount is None:
-                raise ValueError("cumulativeAmount required for open action")
-
-            from mpay.methods.tempo.stream.types import Voucher
-
-            voucher = Voucher(
-                channel_id=channel_id,
-                cumulative_amount=int(cumulative_amount),
-            )
-            signature = sign_voucher(account, voucher, escrow, chain_id)
-            return {
-                "action": "open",
-                "type": "transaction",
-                "channelId": channel_id,
-                "transaction": transaction,
-                "authorizedSigner": context.get("authorizedSigner", account.address),
-                "cumulativeAmount": str(cumulative_amount),
-                "signature": signature,
-            }
-
-        elif action == "topUp":
-            transaction = context.get("transaction")
-            additional_deposit = context.get("additionalDeposit")
-            if not transaction:
-                raise ValueError("transaction required for topUp action")
-            if additional_deposit is None:
-                raise ValueError("additionalDeposit required for topUp action")
-            return {
-                "action": "topUp",
-                "type": "transaction",
-                "channelId": channel_id,
-                "transaction": transaction,
-                "additionalDeposit": str(additional_deposit),
-            }
-
-        elif action == "voucher":
-            cumulative_amount = context.get("cumulativeAmount")
-            if cumulative_amount is None:
-                raise ValueError("cumulativeAmount required for voucher action")
-            return await self._voucher_payload(
-                account,
-                channel_id,
-                int(cumulative_amount),
-                escrow,
-                chain_id,
-            )
-
-        elif action == "close":
-            cumulative_amount = context.get("cumulativeAmount")
-            if cumulative_amount is None:
-                raise ValueError("cumulativeAmount required for close action")
-
-            from mpay.methods.tempo.stream.types import Voucher
-
-            voucher = Voucher(
-                channel_id=channel_id,
-                cumulative_amount=int(cumulative_amount),
-            )
-            signature = sign_voucher(account, voucher, escrow, chain_id)
-            return {
-                "action": "close",
-                "channelId": channel_id,
-                "cumulativeAmount": str(cumulative_amount),
-                "signature": signature,
-            }
-
-        else:
-            raise ValueError(f"unknown action: {action}")
-
-    # ──────────────────────────────────────────────────────────
-    # Helpers
-    # ──────────────────────────────────────────────────────────
-
-    async def _voucher_payload(
-        self,
-        account: TempoAccount,
-        channel_id: str,
-        cumulative_amount: int,
-        escrow_contract: str,
-        chain_id: int,
-    ) -> dict[str, Any]:
-        """Build a voucher credential payload."""
-        from mpay.methods.tempo.stream.types import Voucher
-
-        voucher = Voucher(
-            channel_id=channel_id,
-            cumulative_amount=cumulative_amount,
-        )
-        signature = sign_voucher(account, voucher, escrow_contract, chain_id)
-        return {
-            "action": "voucher",
-            "channelId": channel_id,
-            "cumulativeAmount": str(cumulative_amount),
-            "signature": signature,
-        }
-
-    def _channel_key(self, payee: str, currency: str, escrow: str) -> str:
-        """Generate a cache key for a (payee, currency, escrow) triple."""
-        return f"{payee.lower()}:{currency.lower()}:{escrow.lower()}"
-
-    def _get_method_details(self, challenge: Challenge) -> dict[str, Any]:
-        """Extract methodDetails from a challenge."""
-        md = challenge.request.get("methodDetails", {})
-        return md if isinstance(md, dict) else {}
-
-    def _resolve_chain_id(self, challenge: Challenge) -> int:
-        md = self._get_method_details(challenge)
-        return md.get("chainId", 42431)
-
-    def _resolve_escrow(
-        self,
-        challenge: Challenge,
-        chain_id: int,
-        channel_id: str | None = None,
-    ) -> str:
-        """Resolve the escrow contract address."""
-        if channel_id:
-            cached = self._escrow_map.get(channel_id)
-            if cached:
-                return cached
-
-        md = self._get_method_details(challenge)
-        challenge_escrow = md.get("escrowContract")
-        if challenge_escrow:
-            return challenge_escrow
-
-        if self.escrow_contract:
-            return self.escrow_contract
-
-        defaults = {
-            4217: "0x9d136eEa063eDE5418A6BC7bEafF009bBb6CFa70",
-            42431: "0x9d136eEa063eDE5418A6BC7bEafF009bBb6CFa70",
-        }
-        escrow = defaults.get(chain_id)
-        if escrow:
-            return escrow
-
-        raise ValueError(
-            "No escrowContract available. Provide it in parameters "
-            "or ensure the server challenge includes it."
-        )
-
-
-# ──────────────────────────────────────────────────────────────────
 # Factory
 # ──────────────────────────────────────────────────────────────────
 
@@ -633,7 +202,7 @@ def tempo(
     """Create a Tempo payment method.
 
     Args:
-        intents: Intents to register (e.g. charge, stream).
+        intents: Intents to register (e.g. charge).
         account: Account for signing transactions.
         rpc_url: Tempo RPC endpoint URL.
         root_account: Root account address for access key signing.
