+++ src/mpay/methods/tempo/intents.py
@@ -1,12 +1,11 @@
 """Tempo payment intents (server-side verification).
 
-Implements the charge and stream intents for Tempo payments.
+Implements the charge intent for Tempo payments.
 """
 
 from __future__ import annotations
 
 import asyncio
-from dataclasses import dataclass, replace
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING, Any
 
@@ -18,38 +17,10 @@ from mpay.methods.tempo.schemas import (
     HashCredentialPayload,
     TransactionCredentialPayload,
 )
-from mpay.methods.tempo.stream.chain import (
-    broadcast_open_transaction,
-    broadcast_top_up_transaction,
-    close_on_chain,
-    get_on_chain_channel,
-    settle_on_chain,
-)
-from mpay.methods.tempo.stream.errors import (
-    AmountExceedsDepositError,
-    ChannelClosedError,
-    ChannelConflictError,
-    ChannelNotFoundError,
-    DeltaTooSmallError,
-    InsufficientBalanceError,
-    InvalidSignatureError,
-    StreamError,
-)
-from mpay.methods.tempo.stream.receipt import create_stream_receipt
-from mpay.methods.tempo.stream.storage import ChannelState, SessionState
-from mpay.methods.tempo.stream.voucher import (
-    parse_voucher_from_payload,
-    verify_voucher,
-)
 from mpay.server.intent import VerificationError
 
 if TYPE_CHECKING:
     import httpx
-
-    from mpay.methods.tempo.account import TempoAccount
-    from mpay.methods.tempo.stream.chain import OnChainChannel
-    from mpay.methods.tempo.stream.storage import ChannelStorage
-    from mpay.methods.tempo.stream.types import SignedVoucher, StreamReceipt
 
 
 DEFAULT_TIMEOUT = 30.0
@@ -487,729 +458,3 @@ class ChargeIntent:
             return
 
         raise VerificationError("Invalid transaction: no matching payment call found")
-
-
-# ──────────────────────────────────────────────────────────────────
-# Stream intent
-# ──────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class StreamIntent:
-    """Server-side stream payment intent.
-
-    Verifies stream credentials and manages channel/session state.
-
-    Example::
-
-        from mpay.methods.tempo.stream.storage import MemoryStorage
-        from mpay.methods.tempo import StreamIntent
-
-        storage = MemoryStorage()
-        intent = StreamIntent(
-            storage=storage,
-            rpc_url="https://rpc.tempo.xyz",
-            escrow_contract="0x9d136eEa063eDE5418A6BC7bEafF009bBb6CFa70",
-        )
-    """
-
-    name: str = "stream"
-
-    storage: ChannelStorage = None  # type: ignore[assignment]
-    rpc_url: str = RPC_URL
-    escrow_contract: str = ""
-    chain_id: int = 42431
-    min_voucher_delta: int = 0
-    fee_payer: TempoAccount | None = None
-    fee_payer_url: str = ""  # Sponsor URL for fee-payer-signed transactions
-    account: TempoAccount | None = None  # Server account for settle/close
-
-    def __post_init__(self) -> None:
-        if self.storage is None:
-            raise ValueError(
-                "storage is required for StreamIntent. "
-                "Use MemoryStorage() for single-process servers."
-            )
-
-    async def verify(
-        self,
-        credential: Credential,
-        request: dict[str, Any],
-    ) -> Receipt:
-        """Verify a stream credential.
-
-        Dispatches to the appropriate handler based on the credential
-        payload's ``action`` field.
-
-        Args:
-            credential: The credential to verify.
-            request: The challenge request parameters.
-
-        Returns:
-            A Receipt containing the StreamReceipt data.
-
-        Raises:
-            StreamError: On verification failure.
-        """
-        payload = credential.payload
-        action = payload.get("action")
-
-        challenge = credential.challenge
-        method_details = _get_method_details(challenge, self)
-
-        resolved_fee_payer_url: str | None = None
-        if method_details.get("feePayer") and (self.fee_payer is not None or self.fee_payer_url):
-            resolved_fee_payer_url = self.fee_payer_url or DEFAULT_FEE_PAYER_URL
-        effective_min_delta = (
-            int(method_details["minVoucherDelta"])
-            if method_details.get("minVoucherDelta")
-            else self.min_voucher_delta
-        )
-
-        if action == "open":
-            stream_receipt = await _handle_open(
-                self.storage,
-                self.rpc_url,
-                challenge,
-                payload,
-                method_details,
-                resolved_fee_payer_url,
-                request,
-            )
-        elif action == "topUp":
-            stream_receipt = await _handle_top_up(
-                self.storage,
-                self.rpc_url,
-                challenge,
-                payload,
-                method_details,
-                resolved_fee_payer_url,
-            )
-        elif action == "voucher":
-            stream_receipt = await _handle_voucher(
-                self.storage,
-                self.rpc_url,
-                effective_min_delta,
-                challenge,
-                payload,
-                method_details,
-            )
-        elif action == "close":
-            stream_receipt = await _handle_close(
-                self.storage,
-                self.rpc_url,
-                challenge,
-                payload,
-                method_details,
-                self.account,
-            )
-        else:
-            raise StreamError(f"unknown action: {action}")
-
-        return Receipt(
-            status="success",
-            timestamp=datetime.now(UTC),
-            reference=stream_receipt.channel_id,
-            extra=stream_receipt.to_dict(),
-        )
-
-
-def _get_method_details(
-    challenge: Any,
-    intent: StreamIntent,
-) -> dict[str, Any]:
-    """Extract methodDetails from challenge, falling back to intent."""
-    if hasattr(challenge, "request"):
-        req = challenge.request if isinstance(challenge.request, dict) else {}
-    elif isinstance(challenge, dict):
-        req = challenge.get("request", {})
-    else:
-        req = {}
-
-    md = req.get("methodDetails", {})
-    if not isinstance(md, dict):
-        md = {}
-
-    return {
-        "escrowContract": md.get("escrowContract", intent.escrow_contract),
-        "chainId": md.get("chainId", intent.chain_id),
-        "channelId": md.get("channelId"),
-        "minVoucherDelta": md.get("minVoucherDelta"),
-        "feePayer": md.get("feePayer"),
-    }
-
-
-# ──────────────────────────────────────────────────────────────────
-# Stream public helpers: charge() and settle()
-# ──────────────────────────────────────────────────────────────────
-
-
-async def charge(
-    storage: ChannelStorage,
-    challenge_id: str,
-    amount: int,
-) -> SessionState:
-    """Charge against an active session's balance.
-
-    Args:
-        storage: Channel storage backend.
-        challenge_id: The session/challenge ID.
-        amount: Amount to charge (base units).
-
-    Returns:
-        Updated session state.
-
-    Raises:
-        InsufficientBalanceError: If available balance < amount.
-        ChannelClosedError: If session not found.
-    """
-    session = await storage.update_session(
-        challenge_id,
-        lambda current: _charge_update(current, amount),
-    )
-    if session is None:
-        raise ChannelClosedError("session not found")
-    return session
-
-
-def _charge_update(current: SessionState | None, amount: int) -> SessionState | None:
-    if current is None:
-        return None
-    available = current.accepted_cumulative - current.spent
-    if available < amount:
-        raise InsufficientBalanceError(f"requested {amount}, available {available}")
-    return replace(current, spent=current.spent + amount, units=current.units + 1)
-
-
-async def settle(
-    storage: ChannelStorage,
-    rpc_url: str,
-    escrow_contract: str,
-    channel_id: str,
-    account: TempoAccount,
-) -> str:
-    """One-shot settle: submit highest voucher on-chain.
-
-    Args:
-        storage: Channel storage backend.
-        rpc_url: Tempo RPC endpoint.
-        escrow_contract: Escrow contract address.
-        channel_id: Channel to settle.
-        account: Server account for signing the settle transaction.
-
-    Returns:
-        Transaction hash.
-    """
-    channel = await storage.get_channel(channel_id)
-    if channel is None:
-        raise ChannelNotFoundError("channel not found")
-    if channel.highest_voucher is None:
-        raise StreamError("no voucher to settle")
-
-    settled_amount = channel.highest_voucher.cumulative_amount
-    tx_hash = await settle_on_chain(rpc_url, escrow_contract, channel.highest_voucher, account)
-
-    await storage.update_channel(
-        channel_id,
-        lambda current: _settle_update(current, settled_amount),
-    )
-    return tx_hash
-
-
-def _settle_update(current: ChannelState | None, settled_amount: int) -> ChannelState | None:
-    if current is None:
-        return None
-    next_settled = max(settled_amount, current.settled_on_chain)
-    return replace(current, settled_on_chain=next_settled)
-
-
-# ──────────────────────────────────────────────────────────────────
-# Stream shared voucher verification
-# ──────────────────────────────────────────────────────────────────
-
-
-def _validate_voucher(
-    voucher: SignedVoucher,
-    on_chain: OnChainChannel,
-    authorized_signer: str,
-    method_details: dict[str, Any],
-) -> None:
-    """Validate voucher bounds and signature.
-
-    Raises:
-        ChannelClosedError: If the channel is finalized.
-        StreamError: If voucher amount is below settled amount.
-        AmountExceedsDepositError: If voucher amount exceeds deposit.
-        InvalidSignatureError: If the EIP-712 signature doesn't match.
-    """
-    if on_chain.finalized:
-        raise ChannelClosedError("channel is finalized on-chain")
-    if voucher.cumulative_amount < on_chain.settled:
-        raise StreamError("voucher cumulativeAmount is below on-chain settled amount")
-    if voucher.cumulative_amount > on_chain.deposit:
-        raise AmountExceedsDepositError("voucher amount exceeds on-chain deposit")
-    is_valid = verify_voucher(
-        method_details["escrowContract"],
-        method_details["chainId"],
-        voucher,
-        authorized_signer,
-    )
-    if not is_valid:
-        raise InvalidSignatureError("invalid voucher signature")
-
-
-async def _accept_voucher(
-    storage: ChannelStorage,
-    challenge_id: str,
-    channel_id: str,
-    accepted_cumulative: int,
-) -> SessionState | None:
-    """Atomically upsert a session with a new acceptedCumulative.
-
-    Safe under concurrent requests: cumulative semantics mean the highest
-    acceptedCumulative always wins.
-    """
-
-    def update(existing: SessionState | None) -> SessionState | None:
-        base = existing or SessionState(
-            challenge_id=challenge_id,
-            channel_id=channel_id,
-            accepted_cumulative=0,
-            spent=0,
-            units=0,
-            created_at=datetime.now(UTC),
-        )
-        next_accepted = max(accepted_cumulative, base.accepted_cumulative)
-        return replace(base, accepted_cumulative=next_accepted)
-
-    return await storage.update_session(challenge_id, update)
-
-
-def _validate_on_chain_channel(
-    on_chain: OnChainChannel,
-    recipient: str,
-    currency: str,
-    amount: int | None = None,
-) -> None:
-    """Validate on-chain channel state matches server expectations."""
-    if on_chain.deposit == 0:
-        raise ChannelNotFoundError("channel not funded on-chain")
-    if on_chain.finalized:
-        raise ChannelClosedError("channel is finalized on-chain")
-    if on_chain.payee.lower() != recipient.lower():
-        raise StreamError("on-chain payee does not match server destination")
-    if on_chain.token.lower() != currency.lower():
-        raise StreamError("on-chain token does not match server token")
-    if amount is not None and on_chain.deposit - on_chain.settled < amount:
-        raise InsufficientBalanceError(
-            "channel available balance insufficient for requested amount"
-        )
-
-
-async def _verify_and_accept_voucher(
-    storage: ChannelStorage,
-    rpc_url: str,
-    min_voucher_delta: int,
-    challenge: Any,
-    channel: ChannelState,
-    channel_id: str,
-    voucher: SignedVoucher,
-    on_chain: OnChainChannel,
-    method_details: dict[str, Any],
-) -> StreamReceipt:
-    """Shared logic for verifying an incremental voucher.
-
-    Updates channel state and creates/updates the session.
-    """
-    challenge_id = challenge.id if hasattr(challenge, "id") else challenge["id"]
-
-    if on_chain.finalized:
-        raise ChannelClosedError("channel is finalized on-chain")
-
-    if voucher.cumulative_amount < on_chain.settled:
-        raise StreamError("voucher cumulativeAmount is below on-chain settled amount")
-
-    if voucher.cumulative_amount > on_chain.deposit:
-        raise AmountExceedsDepositError("voucher amount exceeds on-chain deposit")
-
-    # Non-increasing voucher -> idempotent: return current highest
-    if voucher.cumulative_amount <= channel.highest_voucher_amount:
-        session = await _accept_voucher(
-            storage,
-            challenge_id,
-            channel_id,
-            channel.highest_voucher_amount,
-        )
-        if session is None:
-            raise StreamError("failed to create session")
-        return create_stream_receipt(
-            challenge_id=challenge_id,
-            channel_id=channel_id,
-            accepted_cumulative=channel.highest_voucher_amount,
-            spent=session.spent,
-            units=session.units,
-        )
-
-    delta = voucher.cumulative_amount - channel.highest_voucher_amount
-    if delta < min_voucher_delta:
-        raise DeltaTooSmallError(f"voucher delta {delta} below minimum {min_voucher_delta}")
-
-    _validate_voucher(voucher, on_chain, channel.authorized_signer, method_details)
-
-    # Update channel with new highest voucher (atomic)
-    await storage.update_channel(
-        channel_id,
-        lambda current: _update_highest_voucher(current, voucher, on_chain.deposit),
-    )
-
-    session = await _accept_voucher(storage, challenge_id, channel_id, voucher.cumulative_amount)
-    if session is None:
-        raise StreamError("failed to create session")
-
-    return create_stream_receipt(
-        challenge_id=challenge_id,
-        channel_id=channel_id,
-        accepted_cumulative=voucher.cumulative_amount,
-        spent=session.spent,
-        units=session.units,
-    )
-
-
-def _update_highest_voucher(
-    current: ChannelState | None,
-    voucher: SignedVoucher,
-    deposit: int,
-) -> ChannelState | None:
-    if current is None:
-        raise ChannelNotFoundError("channel not found")
-    if voucher.cumulative_amount > current.highest_voucher_amount:
-        return replace(
-            current,
-            deposit=deposit,
-            highest_voucher_amount=voucher.cumulative_amount,
-            highest_voucher=voucher,
-        )
-    return current
-
-
-# ──────────────────────────────────────────────────────────────────
-# Stream action handlers
-# ──────────────────────────────────────────────────────────────────
-
-
-async def _handle_open(
-    storage: ChannelStorage,
-    rpc_url: str,
-    challenge: Any,
-    payload: dict[str, Any],
-    method_details: dict[str, Any],
-    fee_payer_url: str | None,
-    request: dict[str, Any],
-) -> StreamReceipt:
-    """Handle 'open' action: broadcast open tx, verify voucher, create channel."""
-    challenge_id = challenge.id if hasattr(challenge, "id") else challenge["id"]
-    channel_id = payload["channelId"]
-
-    voucher = parse_voucher_from_payload(
-        channel_id, payload["cumulativeAmount"], payload["signature"]
-    )
-
-    # Get recipient and currency from challenge request.
-    # challenge.request may be a base64 string (ChallengeEcho) or a dict.
-    if hasattr(challenge, "request") and isinstance(challenge.request, dict):
-        ch_request = challenge.request
-    elif isinstance(challenge, dict):
-        ch_request = challenge.get("request", {})
-    else:
-        ch_request = {}
-    recipient = ch_request.get("recipient") or request.get("recipient")
-    currency = ch_request.get("currency") or request.get("currency")
-    if not recipient:
-        raise StreamError("could not resolve recipient from challenge")
-    if not currency:
-        raise StreamError("could not resolve currency from challenge")
-    amount_str = ch_request.get("amount", request.get("amount"))
-    amount = int(amount_str) if amount_str else None
-
-    result = await broadcast_open_transaction(
-        rpc_url=rpc_url,
-        serialized_transaction=payload["transaction"],
-        escrow_contract=method_details["escrowContract"],
-        channel_id=channel_id,
-        recipient=recipient,
-        currency=currency,
-        fee_payer_url=fee_payer_url,
-    )
-    on_chain = result.on_chain
-
-    _validate_on_chain_channel(on_chain, recipient, currency, amount)
-
-    # Resolve authorized signer: zero address -> use payer
-    authorized_signer = (
-        on_chain.payer if on_chain.authorized_signer == ZERO_ADDRESS else on_chain.authorized_signer
-    )
-
-    _validate_voucher(voucher, on_chain, authorized_signer, method_details)
-
-    session = await _accept_voucher(storage, challenge_id, channel_id, voucher.cumulative_amount)
-    if session is None:
-        raise StreamError("failed to create session")
-
-    existing_channel = await storage.get_channel(channel_id)
-    stale_session_id: str | None = None
-    if existing_channel and existing_channel.active_session_id:
-        active_session = await storage.get_session(existing_channel.active_session_id)
-        if active_session is None:
-            stale_session_id = existing_channel.active_session_id
-
-    try:
-        await storage.update_channel(
-            channel_id,
-            lambda existing: _open_channel_update(
-                existing,
-                challenge_id,
-                stale_session_id,
-                voucher,
-                on_chain,
-                authorized_signer,
-            ),
-        )
-    except Exception:
-        # Clean up pre-created session on conflict/failure
-        await storage.update_session(challenge_id, lambda _: None)
-        raise
-
-    return create_stream_receipt(
-        challenge_id=challenge_id,
-        channel_id=channel_id,
-        accepted_cumulative=voucher.cumulative_amount,
-        spent=session.spent,
-        units=session.units,
-    )
-
-
-def _open_channel_update(
-    existing: ChannelState | None,
-    challenge_id: str,
-    stale_session_id: str | None,
-    voucher: SignedVoucher,
-    on_chain: OnChainChannel,
-    authorized_signer: str,
-) -> ChannelState:
-    if existing is not None:
-        # Check for concurrent stream
-        if (
-            existing.active_session_id
-            and existing.active_session_id != challenge_id
-            and existing.active_session_id != stale_session_id
-        ):
-            raise ChannelConflictError("another stream is active on this channel")
-
-        if voucher.cumulative_amount < existing.settled_on_chain:
-            raise StreamError("voucher amount is below settled on-chain amount")
-
-        if voucher.cumulative_amount > existing.highest_voucher_amount:
-            return replace(
-                existing,
-                deposit=on_chain.deposit,
-                highest_voucher_amount=voucher.cumulative_amount,
-                highest_voucher=voucher,
-                authorized_signer=authorized_signer,
-                active_session_id=challenge_id,
-            )
-        return replace(
-            existing,
-            deposit=on_chain.deposit,
-            authorized_signer=authorized_signer,
-            active_session_id=challenge_id,
-        )
-
-    return ChannelState(
-        channel_id=voucher.channel_id,
-        payer=on_chain.payer,
-        payee=on_chain.payee,
-        token=on_chain.token,
-        authorized_signer=authorized_signer,
-        deposit=on_chain.deposit,
-        settled_on_chain=0,
-        highest_voucher_amount=voucher.cumulative_amount,
-        highest_voucher=voucher,
-        active_session_id=challenge_id,
-        finalized=False,
-        created_at=datetime.now(UTC),
-    )
-
-
-async def _handle_top_up(
-    storage: ChannelStorage,
-    rpc_url: str,
-    challenge: Any,
-    payload: dict[str, Any],
-    method_details: dict[str, Any],
-    fee_payer_url: str | None,
-) -> StreamReceipt:
-    """Handle 'topUp' action: broadcast topUp tx, update deposit.
-
-    Per spec Section 8.3.2, topUp payloads contain only the transaction
-    and additionalDeposit -- no voucher.
-    """
-    challenge_id = challenge.id if hasattr(challenge, "id") else challenge["id"]
-    channel_id = payload["channelId"]
-
-    channel = await storage.get_channel(channel_id)
-    if channel is None:
-        raise ChannelNotFoundError("channel not found")
-
-    declared_deposit = int(payload["additionalDeposit"])
-
-    _tx_hash, on_chain_deposit = await broadcast_top_up_transaction(
-        rpc_url=rpc_url,
-        serialized_transaction=payload["transaction"],
-        escrow_contract=method_details["escrowContract"],
-        channel_id=channel_id,
-        declared_deposit=declared_deposit,
-        previous_deposit=channel.deposit,
-        fee_payer_url=fee_payer_url,
-    )
-
-    await storage.update_channel(
-        channel_id,
-        lambda current: replace(current, deposit=on_chain_deposit) if current else None,
-    )
-
-    session = await storage.get_session(challenge_id)
-
-    return create_stream_receipt(
-        challenge_id=challenge_id,
-        channel_id=channel_id,
-        accepted_cumulative=(
-            session.accepted_cumulative if session else channel.highest_voucher_amount
-        ),
-        spent=session.spent if session else 0,
-        units=session.units if session else 0,
-    )
-
-
-async def _handle_voucher(
-    storage: ChannelStorage,
-    rpc_url: str,
-    min_voucher_delta: int,
-    challenge: Any,
-    payload: dict[str, Any],
-    method_details: dict[str, Any],
-) -> StreamReceipt:
-    """Handle 'voucher' action: verify and accept a new voucher."""
-    channel_id = payload["channelId"]
-
-    channel = await storage.get_channel(channel_id)
-    if channel is None:
-        raise ChannelNotFoundError("channel not found")
-    if channel.finalized:
-        raise ChannelClosedError("channel is finalized")
-
-    voucher = parse_voucher_from_payload(
-        channel_id, payload["cumulativeAmount"], payload["signature"]
-    )
-
-    on_chain = await get_on_chain_channel(rpc_url, method_details["escrowContract"], channel_id)
-
-    return await _verify_and_accept_voucher(
-        storage=storage,
-        rpc_url=rpc_url,
-        min_voucher_delta=min_voucher_delta,
-        challenge=challenge,
-        channel=channel,
-        channel_id=channel_id,
-        voucher=voucher,
-        on_chain=on_chain,
-        method_details=method_details,
-    )
-
-
-async def _handle_close(
-    storage: ChannelStorage,
-    rpc_url: str,
-    challenge: Any,
-    payload: dict[str, Any],
-    method_details: dict[str, Any],
-    server_account: TempoAccount | None,
-) -> StreamReceipt:
-    """Handle 'close' action: verify final voucher, close channel."""
-    challenge_id = challenge.id if hasattr(challenge, "id") else challenge["id"]
-    channel_id = payload["channelId"]
-
-    channel = await storage.get_channel(channel_id)
-    if channel is None:
-        raise ChannelNotFoundError("channel not found")
-    if channel.finalized:
-        raise ChannelClosedError("channel is already finalized")
-
-    voucher = parse_voucher_from_payload(
-        channel_id, payload["cumulativeAmount"], payload["signature"]
-    )
-
-    if voucher.cumulative_amount < channel.highest_voucher_amount:
-        raise StreamError("close voucher amount must be >= highest accepted voucher")
-
-    on_chain = await get_on_chain_channel(rpc_url, method_details["escrowContract"], channel_id)
-
-    if on_chain.finalized:
-        raise ChannelClosedError("channel is finalized on-chain")
-
-    if voucher.cumulative_amount < on_chain.settled:
-        raise StreamError("close voucher cumulativeAmount is below on-chain settled amount")
-
-    if voucher.cumulative_amount > on_chain.deposit:
-        raise AmountExceedsDepositError("close voucher amount exceeds on-chain deposit")
-
-    is_valid = verify_voucher(
-        method_details["escrowContract"],
-        method_details["chainId"],
-        voucher,
-        channel.authorized_signer,
-    )
-    if not is_valid:
-        raise InvalidSignatureError("invalid voucher signature")
-
-    session = await storage.get_session(challenge_id)
-
-    tx_hash: str | None = None
-    if server_account is not None:
-        tx_hash = await close_on_chain(
-            rpc_url,
-            method_details["escrowContract"],
-            voucher,
-            server_account,
-        )
-
-    await storage.update_channel(
-        channel_id,
-        lambda current: _close_channel_update(current, voucher, on_chain.deposit),
-    )
-    await storage.update_session(challenge_id, lambda _: None)
-
-    return create_stream_receipt(
-        challenge_id=challenge_id,
-        channel_id=channel_id,
-        accepted_cumulative=voucher.cumulative_amount,
-        spent=session.spent if session else 0,
-        units=session.units if session else 0,
-        tx_hash=tx_hash,
-    )
-
-
-def _close_channel_update(
-    current: ChannelState | None,
-    voucher: SignedVoucher,
-    deposit: int,
-) -> ChannelState | None:
-    if current is None:
-        return None
-    return replace(
-        current,
-        deposit=deposit,
-        highest_voucher_amount=voucher.cumulative_amount,
-        highest_voucher=voucher,
-        active_session_id=None,
-        finalized=True,
-    )
