--- src/mpay/methods/tempo/stream/__init__.py
@@ -1,5 +1,0 @@
-"""Tempo stream payment intent — pay-as-you-go payment channels.
-
-Uses cumulative vouchers over an on-chain escrow contract for
-incremental micropayments without per-request transactions.
-"""
