+++ src/mpay/server/mpay.py
@@ -146,67 +146,3 @@ class Mpay:
             method=self.method.name,
             description=description,
         )
-
-    async def stream(
-        self,
-        authorization: str | None,
-        amount: str,
-        *,
-        unit_type: str = "token",
-        currency: str | None = None,
-        recipient: str | None = None,
-        description: str | None = None,
-    ) -> Challenge | tuple[Credential, Receipt]:
-        """Handle a stream intent.
-
-        Args:
-            authorization: The Authorization header value (or None).
-            amount: Price per unit in human units (e.g., "0.000075").
-                Automatically converted to base units.
-            unit_type: Service unit type (e.g., "token", "byte").
-            currency: Override the method's default currency.
-            recipient: Override the method's default recipient.
-            description: Optional human-readable description.
-
-        Returns:
-            Challenge if payment required, or (Credential, Receipt) if verified.
-        """
-        intent = self.method.intents.get("stream")
-        if intent is None:
-            raise ValueError(f"Method {self.method.name} does not support stream intent")
-
-        resolved_currency = currency or getattr(self.method, "currency", None)
-        resolved_recipient = recipient or getattr(self.method, "recipient", None)
-        if not resolved_currency:
-            raise ValueError("currency must be set on the method or passed to stream()")
-        if not resolved_recipient:
-            raise ValueError("recipient must be set on the method or passed to stream()")
-
-        decimals = getattr(self.method, "decimals", DEFAULT_DECIMALS)
-        base_amount = str(parse_units(amount, decimals))
-
-        escrow_contract = getattr(self.method, "escrow_contract", None) or getattr(
-            intent, "escrow_contract", ""
-        )
-        chain_id = getattr(self.method, "chain_id", None) or getattr(intent, "chain_id", 42431)
-
-        request: dict[str, Any] = {
-            "amount": base_amount,
-            "unitType": unit_type,
-            "currency": resolved_currency,
-            "recipient": resolved_recipient,
-            "methodDetails": {
-                "escrowContract": escrow_contract,
-                "chainId": chain_id,
-            },
-        }
-
-        return await verify_or_challenge(
-            authorization=authorization,
-            intent=intent,
-            request=request,
-            realm=self.realm,
-            secret_key=self.secret_key,
-            method=self.method.name,
-            description=description,
-        )
